add book dto  fix

+++ src/main/java/team2/spring/library/dto/GeneralStatisticDto.java
@@ -1,6 +1,5 @@
 package team2.spring.library.dto;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,8 +8,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class GeneralStatisticDto {
-    private  long avgDaysOfReading;
-    private int avgAgeOfReaders;
-   private  int avgVisitOfLibrary;
-
+  private long avgDaysOfReading;
+  private double avgAgeOfReaders;
+  private int avgVisitOfLibrary;
 }
